+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ### Version 0.0.2.1
 * (work in progress)
+* Add main loop for progress tracking, and add node data to persistence file.
 * Add game setting 'Track Progress'.
 
 ### Version 0.0.1.3